% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\name{goodbye}
\alias{goodbye}
\title{Say Goodbye}
\usage{
goodbye(name = "Fred")
}
\arguments{
\item{name}{The name of a person that is leaving. Perhaps they are taking a vacation}
}
\description{
This is a simple function with a difficult execution, unless you are not fond of the individual.
}
\examples{
goodbye()
goodbye('Hal')
}
