% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello(name = "Fred")
}
\arguments{
\item{name}{The name of a person to be used in the greeting.}
}
\description{
This is a pretty simple function to extend greetings.
}
\examples{
hello()
hello("Kitty")
}
