% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thanks.R
\name{thanks}
\alias{thanks}
\title{Say thanks}
\usage{
thanks(name = "Fred")
}
\arguments{
\item{name}{The name of a person to be thanked}
}
\description{
This is a function to automate kindness
That is the only description needed
}
